BOXED style independent from ax object
