+++ flowmapviz/__init__.py
@@ -1,1 +1,3 @@
-from .plot import plot_routes, WidthStyle, map_distance_to_point_units
+from .plot import plot_routes, WidthStyle
+from .preprocessing import map_distance_to_point_units
+from .zoom import get_zoom_level, ZoomLevel, plot_graph_with_zoom
