+++ flowmapviz/plot.py
@@ -29,16 +29,6 @@ class WidthStyle(Enum):
     """
 
 
-def map_distance_to_point_units(map_distance: float, ax):
-    lims = np.array([lim[1] - lim[0] for lim in (ax.get_xlim(), ax.get_ylim())])
-    return map_distance * ax.get_window_extent().size / lims
-
-
-def point_units_to_map_distance(value_in_points: float, ax):
-    lims = np.array([lim[1] - lim[0] for lim in (ax.get_xlim(), ax.get_ylim())])
-    return value_in_points * lims / ax.get_window_extent().size
-
-
 def plot_routes(g: nx.MultiDiGraph,
                 ax: Axes,
                 nodes_from: list[int],
@@ -66,7 +56,7 @@ def plot_routes(g: nx.MultiDiGraph,
     :param width_style: style of the width representation
     :param round_edges: if True plot circles at the end of wide segments for smoother connection
     :param plot: if True add collections to Ax
-    :param filter_by_zoom: if True filter segments based on the zoom level of ax
+    :param roadtypes_by_zoom: if True filter segments based on the zoom level of ax
     :return: LineCollection of color segments, PatchCollection of width representation
     """
     lines = []
@@ -156,7 +146,6 @@ def plot_route(g: nx.MultiDiGraph,
     color_scalar = np.interp(density_index, np.arange(len(densities)), densities)
 
     # width as filling
-    width_modifier, wm2 = point_units_to_map_distance(width_modifier, ax)
     polygons = []
     if width_style == WidthStyle.CALLIGRAPHY:
         polygons = get_width_polygon(ax, x, y, point_densities, min_width_density, max_width_density,
